// src/api/generated/models/TimesheetFolder.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Blackwell Time API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TimesheetFolder
 */
export interface TimesheetFolder {
    /**
     * 
     * @type {string}
     * @memberof TimesheetFolder
     */
    timesheetFolderId?: string;
    /**
     * 
     * @type {string}
     * @memberof TimesheetFolder
     */
    timesheetFolderName?: string;
    /**
     * 
     * @type {string}
     * @memberof TimesheetFolder
     */
    driveFolderId?: string;
    /**
     * 
     * @type {TimesheetFolderStatusEnum}
     * @memberof TimesheetFolder
     */
    status?: TimesheetFolderStatusEnum;
}


/**
 * @export
 */
export const TimesheetFolderStatusEnum = {
    Active: 'Active',
    Inactive: 'Inactive'
} as const;
export type TimesheetFolderStatusEnum = typeof TimesheetFolderStatusEnum[keyof typeof TimesheetFolderStatusEnum];


/**
 * Check if a given object implements the TimesheetFolder interface.
 */
export function instanceOfTimesheetFolder(value: object): value is TimesheetFolder {
    return true;
}

export function TimesheetFolderFromJSON(json: any): TimesheetFolder {
    return TimesheetFolderFromJSONTyped(json, false);
}

export function TimesheetFolderFromJSONTyped(json: any, ignoreDiscriminator: boolean): TimesheetFolder {
    if (json == null) {
        return json;
    }
    return {
        
        'timesheetFolderId': json['timesheetFolderId'] == null ? undefined : json['timesheetFolderId'],
        'timesheetFolderName': json['timesheetFolderName'] == null ? undefined : json['timesheetFolderName'],
        'driveFolderId': json['driveFolderId'] == null ? undefined : json['driveFolderId'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function TimesheetFolderToJSON(json: any): TimesheetFolder {
    return TimesheetFolderToJSONTyped(json, false);
}

export function TimesheetFolderToJSONTyped(value?: TimesheetFolder | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'timesheetFolderId': value['timesheetFolderId'],
        'timesheetFolderName': value['timesheetFolderName'],
        'driveFolderId': value['driveFolderId'],
        'status': value['status'],
    };
}